Moved helpers to sequence_utils.py to prepare for radon transform implementation
